move time utils to time.rs
